Handle multi-module projects in Gradle (Java)
Some Java projects contain multiple more or less separate sub-projects.
In such a case we cannot assume that the current working directory is
the root of the sub-project we wish to test. Invoking `gradle test`
would run the test class in all sub-project and fail if any one of the
sub-project does not have a test class with that name.

This change makes an educated guess: if the working directory contains a
directory named `src` we assume that the project only has one module.
Otherwise we use the first component of the file path of the test file
as the name of the sub-project. This guess only works if the directory
structure of the project follows the conventional Gradle guidelines.